Cambios requeridos en la pull request con sus componentes por separados, añadir identificador único a cada task y renombrar componentes ItemList por ListItem
