+++ src/App.tsx
@@ -1,6 +1,6 @@
 
 import './App.css'
-import { ToDoWrapper } from './components/Lists'
+import { ToDoWrapper } from './components/ToDoWrapper'
 
 function App() {
