+++ src/components/AddButton.tsx
@@ -1,0 +1,3 @@
+export function AddButton() {
+    return <button>Add +</button>;
+}
