+++ src/components/Lists.tsx
@@ -1,57 +1,10 @@
-interface ItemProps {
-    title: string;
-    description?: string;
-}
+import { ListItem } from "./ListItem";
 
-interface ToDoProps {
-    title: string;
-}
-
-export function ToDoWrapper({ title }: ToDoProps) {
-    return (
-        <main className="to-do-wrapper">
-            <div className="hero-title">
-                <h1>{title}</h1>
-            </div>
-            <div className="task-wrapper">
-                <div className="buttons-wrapper">
-                    <AddButton />
-                    <FilterButton />
-                </div>
-                <List />
-            </div>
-        </main>
-    );
-}
-
-export function AddButton() {
-    return <button>Add +</button>;
-}
-
-export function FilterButton() {
-    return (
-        <>
-            <select name="options" id="options">
-                <option value="all">All</option>
-                <option value="doing">Doing</option>
-                <option value="done">Done</option>
-            </select>
-        </>
-    );
-}
-
-export function ItemList({ title, description }: ItemProps) {
-    return (
-        <div className="list-item">
-            <h1>{title}</h1>
-            {description && <p>{description}</p>}
-        </div>
-    );
-}
 
 export function List() {
     let data = [
         {
+            id: "1",
             title: "hello",
             description: "This is a description",
             dateCreated: "12/08/2000",
@@ -60,6 +13,7 @@ export function List() {
             doing: "true"
         },
         {
+            id: "2",
             title: "hello",
             description: "This is a description",
             dateCreated: "12/08/2000",
@@ -71,9 +25,9 @@ export function List() {
 
     return (
         <div className="list-wrapper">
-            {data.map((item, index) => (
-                <ItemList
-                    key={index}
+            {data.map((item) => (
+                <ListItem
+                    key={item.id}
                     title={item.title}
                     description={item.description}
                 />
